<feat> Bootstrap and pagination added

+++ week-08/day-05/reddit/src/main/java/com/greenfox/reddit/model/Post.java
@@ -9,8 +9,8 @@ import javax.persistence.Id;
 public class Post {
 
     @Id
-    @GeneratedValue(strategy= GenerationType.AUTO)
-    private Long id;
+    @GeneratedValue(strategy = GenerationType.AUTO)
+    private Integer id;
     private String content;
     private int score;
 
@@ -28,12 +28,12 @@ public class Post {
 
     }
 
-    public Long getId() {
+    public Integer getId() {
 
         return id;
     }
 
-    public void setId(long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
